Change async fetch to beforeCreate
This change ensures that the appropriate information is retrieved on page load. Otherwise, going to a direct URL (e.g. /posts/23) the async method did not load the information. Now it  does.

However, there is a notable delay before content is rendered. Perhaps add a transition (to hide this delay), or re-work the dispatch function to perform quicker.