Use nuxtServerInit to populate posts on store init
This gets all the posts on the initialization of the store.

This change means that when a user goes to /posts/34, Post 34 is loaded already into the store and we do not have to wait for the dispatch function to retrieve all posts again. This makes the pages load significantly faster as an API request does not trigger on every page load :cake:.